Fix cv2_image_collate_fn.

+++ torchdatapipe/collections/vision/utils/general.py
@@ -23,4 +23,4 @@ def cv2_image_collate_fn(images, norm=255):
     images = np.stack(images)
     images = np.moveaxis(images, [0, 1, 2, 3], [0, 2, 3, 1])
     images = np.flip(images, axis=1)  # BGR -> RGB
-    return images / norm
+    return (images / norm).astype(np.float32)
